% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up.R
\name{constant}
\alias{constant}
\alias{constant.data.frame}
\alias{constant.default}
\title{Generic test for constant}
\usage{
constant(x, ...)

\method{constant}{default}(x, na.rm = FALSE, ...)

\method{constant}{data.frame}(x, id, all = FALSE, ...)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{(default FALSE)}

\item{id}{clustering variable}

\item{all}{all variable must be constant (default FALSE)}

\item{\dots}{not currently used}

\item{\dots}{not currently used}
}
\description{
Generic test for constant
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method for constant

\item \code{data.frame}: data.frame method for constant
}}

