% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{popagetable}
\alias{popagetable}
\title{Population frequencies (CAUTION: Use only for testing)}
\format{A data frame with 450,800 rows and 8 variables.
\describe{
  \item{state.code}{first 2 digits of county code identifying state}
  \item{sex}{Male or Female}
  \item{age}{an ordered factor with 14 age categories}
  \item{raceethnicity}{in 6 categories}
  \item{county}{5-digit county code}
  \item{state}{2-letter state abbreviation}
  \item{country}{US or PR for Puerto Rico}
  \item{population}{count of people in a particular combination (Note that this number has been imputed through some process and should be used only for testing, not for production)}
}}
\source{
Datassist plus some arbitraty adjustments for population
}
\description{
This data frame contains the number of individuals within each county in the 
US and Puerto Rico (PR) in clusters formed by combining categories of 
age (in 14 categories), raceethnicity (in 5 categories) and sex (Male or Female).
Each county is also identified by its 5-digit county code,  
its state (50 states using 2-letter abbreviations plus DC and PR),
and its country (US or PR).
}

